requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
jinja2==3.1.4
python-jose==3.3.0
pyOpenSSL==24.2.1
SQLAlchemy==2.0.35
aiofiles==24.1.0
python-multipart==0.0.12